+++ lib/src/config/smart_config.dart
@@ -1,5 +1,3 @@
-import '../kit/dialog_kit.dart';
-import 'enum_config.dart';
 import 'smart_config_attach.dart';
 import 'smart_config_custom.dart';
 import 'smart_config_loading.dart';
@@ -34,44 +32,4 @@ class SmartConfig {
   ///
   /// showToast(): toast全局配置项
   SmartConfigToast toast = SmartConfigToast();
-
-  /// Check whether the relevant dialog exists on the interface,
-  /// if the tag attribute is used, dialogTypes will be invalid
-  ///
-  /// 检查相关dialog是否存在于界面上，如果使用tag属性，dialogTypes将失效
-  @Deprecated("please use SmartDialog.checkExist()")
-  bool checkExist({
-    String? tag,
-    Set<SmartAllDialogType> dialogTypes = const {
-      SmartAllDialogType.custom,
-      SmartAllDialogType.attach,
-      SmartAllDialogType.loading,
-    },
-  }) {
-    return DialogKit.instance.checkExist(tag: tag, dialogTypes: dialogTypes);
-  }
-
-  /// whether custom dialog，attach and loading  exist on the screen
-  ///
-  /// 自定义dialog，attach或loading，是否存在在界面上
-  @Deprecated("Please use checkExist")
-  bool get isExist => custom.isExist || attach.isExist || loading.isExist;
-
-  /// whether custom dialog exist on the screen
-  ///
-  /// 自定义dialog或attach是否存在在界面上
-  @Deprecated("Please use checkExist")
-  bool get isExistDialog => custom.isExist || attach.isExist;
-
-  /// whether loading exist on the screen
-  ///
-  /// loading是否存在界面上
-  @Deprecated("Please use checkExist")
-  bool get isExistLoading => loading.isExist;
-
-  /// whether toast exist on the screen
-  ///
-  /// toast是否存在在界面上
-  @Deprecated("Please use checkExist")
-  bool get isExistToast => toast.isExist;
 }