+++ lib/src/config/smart_config_custom.dart
@@ -17,7 +17,6 @@ class SmartConfigCustom {
     this.clickMaskDismiss = true,
     this.debounce = false,
     this.debounceTime = const Duration(milliseconds: 300),
-    this.backDismiss = true,
     this.bindPage = true,
     this.awaitOverType = SmartAwaitOverType.dialogDismiss,
     this.maskTriggerType = SmartMaskTriggerType.up,
@@ -104,15 +103,6 @@ class SmartConfigCustom {
   /// [debounceTime]：防抖时间内，多次点击只会响应第一次，第二次无效点击会触发防抖时间重新计时
   /// note: 防抖定义是在规定时间无二次触发才会响应,这样会导致点击响应有延时,所以实际实现做一点调整
   final Duration debounceTime;
-
-  /// true（the back event will close the loading but not close the page），
-  /// false（the back event not close the loading and not close page），
-  /// you still can use the dismiss method to close the loading
-  ///
-  /// true（返回事件将关闭loading，但是不会关闭页面），false（返回事件不会关闭loading，也不会关闭页面），
-  /// 你仍然可以使用dismiss方法来关闭dialog
-  @Deprecated("please use backType")
-  final bool backDismiss;
 
   /// Bind the dialog to the current page, the bound page is not on the top of the stack,
   /// the dialog is automatically hidden, the bound page is placed on the top of the stack, and the dialog is automatically displayed;